Remove inherits package
