+++ benchmark/benchmark.js
@@ -5,10 +5,10 @@ var ProgressBar = require('progress')
 
 var util = require('../test/util')
 var implementations = {
-  // bindings: require('../bindings'),
+  bindings: require('../bindings'),
   secp256k1js: require('../js'),
-  elliptic: require('./elliptic')
-  // ecdsa: require('./ecdsa')
+  elliptic: require('./elliptic'),
+  ecdsa: require('./ecdsa')
 }
 
 var fixtureIndex = 0
