+++ lib/ecpoint.js
@@ -244,67 +244,6 @@ ECPoint.prototype.mul = function (num) {
 }
 
 /**
- * @param {BN} k1
- * @param {ECPoint} p2
- * @param {BN} k2
- * @return {ECPoint}
- */
-ECPoint.prototype.mulAdd = function (k1, p2, k2) {
-  var p1 = this
-
-  var nafPointsP1 = p1._getNAFPoints(1)
-  var nafPointsP2 = p2._getNAFPoints(1)
-  var wnd = [nafPointsP1.points, nafPointsP2.points]
-  var naf = [util.getNAF(k1, nafPointsP1.wnd), util.getNAF(k2, nafPointsP2.wnd)]
-
-  var acc = new ECJPoint(null, null, null)
-  var tmp = new Array(2)
-  for (var i = Math.max(naf[0].length, naf[1].length); i >= 0; i--) {
-    var k = 0
-
-    for (; i >= 0; ++k, --i) {
-      tmp[0] = naf[0][i] | 0
-      tmp[1] = naf[1][i] | 0
-
-      if (tmp[0] !== 0 || tmp[1] !== 0) {
-        break
-      }
-    }
-
-    if (i >= 0) {
-      k++
-    }
-
-    acc = acc.dblp(k)
-
-    if (i < 0) {
-      break
-    }
-
-    for (var jj = 0; jj < 2; jj++) {
-      var z = tmp[jj]
-      var p
-      if (z === 0) {
-        continue
-      } else if (z > 0) {
-        p = wnd[jj][z >> 1]
-      } else if (z < 0) {
-        p = wnd[jj][-z >> 1].neg()
-      }
-
-      // hack: ECPoint detection
-      if (p.z === undefined) {
-        acc = acc.mixedAdd(p)
-      } else {
-        acc = acc.add(p)
-      }
-    }
-  }
-
-  return ECPoint.fromECJPoint(acc)
-}
-
-/**
  * @param {number} wnd
  * @return {{wnd: number, points: ECPoint[]}}
  */
