+++ lib/ecpointg.js
@@ -1,7 +1,5 @@
 'use strict'
 
-var inherits = require('inherits')
-
 var ecparams = require('./ecparams')
 var ECPoint = require('./ecpoint')
 var ECJPoint = require('./ecjpoint')
@@ -9,21 +7,41 @@ var util = require('./util')
 
 /**
  * @class ECPointG
- * @extends ECPoint
  * @constructor
- * @param {BN} gx
- * @param {BN} gy
  */
-function ECPointG (gx, gy) {
-  ECPoint.call(this, gx, gy)
+function ECPointG () {
+  this.x = ecparams.gx
+  this.y = ecparams.gy
+  this.inf = false
 
-  this.precomputed = {
-    doubles: this._getDoubles(4),
-    naf: ECPoint.prototype._getNAFPoints.call(this, 8)
-  }
+  this._precompute()
 }
 
-inherits(ECPointG, ECPoint)
+/**
+ */
+ECPointG.prototype._precompute = function () {
+  var ecpoint = new ECPoint(this.x, this.y)
+
+  var dstep = 4
+  var points = new Array(1 + Math.ceil(257 / dstep))
+  var acc = points[0] = ecpoint
+  for (var i = 1; i < points.length; ++i) {
+    for (var j = 0; j < dstep; j++) {
+      acc = acc.dbl()
+    }
+
+    points[i] = acc
+  }
+
+  this.precomputed = {
+    naf: ecpoint._getNAFPoints(8),
+    doubles: {
+      step: dstep,
+      points: points,
+      negpoints: points.map(function (p) { return p.neg() })
+    }
+  }
+}
 
 /**
  * @param {BN} num
@@ -67,31 +85,62 @@ ECPointG.prototype.mul = function (num) {
 }
 
 /**
- * @param {number} step
- * @return {{step: number, points: ECPointG[]}
+ * @param {BN} k1
+ * @param {ECPoint} p2
+ * @param {BN} k2
+ * @return {ECPoint}
  */
-ECPointG.prototype._getDoubles = function (step) {
-  var points = new Array(1 + Math.ceil(257 / step))
-  points[0] = this
+ECPointG.prototype.mulAdd = function (k1, p2, k2) {
+  var nafPointsP1 = this.precomputed.naf
+  var nafPointsP2 = p2._getNAFPoints(1)
+  var wnd = [nafPointsP1.points, nafPointsP2.points]
+  var naf = [util.getNAF(k1, nafPointsP1.wnd), util.getNAF(k2, nafPointsP2.wnd)]
 
-  var acc = this
-  for (var i = 1; i < points.length; ++i) {
-    for (var j = 0; j < step; j++) {
-      acc = acc.dbl()
+  var acc = new ECJPoint(null, null, null)
+  var tmp = new Array(2)
+  for (var i = Math.max(naf[0].length, naf[1].length); i >= 0; i--) {
+    var k = 0
+
+    for (; i >= 0; ++k, --i) {
+      tmp[0] = naf[0][i] | 0
+      tmp[1] = naf[1][i] | 0
+
+      if (tmp[0] !== 0 || tmp[1] !== 0) {
+        break
+      }
     }
 
-    points[i] = acc
+    if (i >= 0) {
+      k++
+    }
+
+    acc = acc.dblp(k)
+
+    if (i < 0) {
+      break
+    }
+
+    for (var jj = 0; jj < 2; jj++) {
+      var z = tmp[jj]
+      var p
+      if (z === 0) {
+        continue
+      } else if (z > 0) {
+        p = wnd[jj][z >> 1]
+      } else if (z < 0) {
+        p = wnd[jj][-z >> 1].neg()
+      }
+
+      // hack: ECPoint detection
+      if (p.z === undefined) {
+        acc = acc.mixedAdd(p)
+      } else {
+        acc = acc.add(p)
+      }
+    }
   }
 
-  var neg = function (p) { return p.neg() }
-  return {step: step, points: points, negpoints: points.map(neg)}
+  return ECPoint.fromECJPoint(acc)
 }
 
-/**
- * @return {{wnd: number, points: ECPointG[]}}
- */
-ECPointG.prototype._getNAFPoints = function () {
-  return this.precomputed.naf
-}
-
-module.exports = new ECPointG(ecparams.gx, ecparams.gy)
+module.exports = new ECPointG()
