+++ package.json
@@ -53,7 +53,6 @@
   "dependencies": {
     "bindings": "^1.2.1",
     "bn.js": "^4.6.2",
-    "inherits": "^2.0.1",
     "nan": "^2.0.9"
   },
   "devDependencies": {
